playground/language-learning-3rd-speaker/requirements.txt
python-dotenv==1.0.0
webrtcvad==2.0.10
pyaudio==0.2.13 